Modify existing instructions to obey the linear bit
In general, for instructions with move semantics, the tag of the
source register is cleared. For instructions with update semantics,
an exception will be raised if the destination register is not
equal to the source register. For store instructions the tag of the
source register is cleared and for load instructions the tag of the
destination register is cleared.